cleaning GUI of login methods
