+++ app/src/main/java/ca/klapstein/baudit/views/LoginView.java
@@ -4,6 +4,7 @@ public interface LoginView extends View {
     void onLoginValidationSuccess();
     void onLoginValidationFailure(String message);
 
+    void startRegistration();
     void startScanQRCode();
     void switchLoginScreen();
 }
