+++ app/src/main/res/layout/activity_login_care_provider.xml
@@ -27,54 +27,22 @@
         android:textSize="24sp"
         android:textStyle="bold" />
 
-    <EditText
-        android:id="@+id/enter_care_provider_username"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="20dp"
-        android:layout_marginRight="20dp"
-        android:ems="10"
-        android:hint="@string/enter_username"
-        android:inputType="text"
-        android:textColor="@color/dark_grey" />
-
-    <EditText
-        android:id="@+id/enter_care_provider_password"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="20dp"
-        android:layout_marginRight="20dp"
-        android:layout_marginTop="16dp"
-        android:ems="10"
-        android:hint="@string/enter_password"
-        android:inputType="textPassword"
-        android:textColor="@color/dark_grey" />
-
     <TextView
-        android:id="@+id/care_provider_login_error_text"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:textColor="@color/red"
-        android:layout_marginRight="20dp"
-        android:layout_marginLeft="20dp" />
+            android:id="@+id/login_error_text"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:textColor="@color/red"
+            android:layout_marginRight="20dp"
+            android:layout_marginLeft="20dp"/>
 
     <Button
-        android:id="@+id/login_care_provider_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="15dp"
-        android:text="@string/log_in"
-        android:textAllCaps="true"
-        android:textSize="18sp" />
-
-    <Button
-        android:id="@+id/register_care_provider_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="14dp"
-        android:text="@string/sign_up"
-        android:textAllCaps="true"
-        android:textSize="14sp"/>
+            android:id="@+id/register_account_button"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="14dp"
+            android:text="@string/sign_up"
+            android:textAllCaps="true"
+            android:textSize="14sp"/>
 
     <Button
             android:id="@+id/scan_qr_code_button"
@@ -84,10 +52,10 @@
             android:textColor="@color/dark_grey"/>
 
     <Button
-        android:id="@+id/log_in_as_patient_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/log_in_as_patient"
-        android:textColor="@android:color/black"/>
+            android:id="@+id/log_in_other_button"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/log_in_as_patient"
+            android:textColor="@android:color/black"/>
 
 </LinearLayout>
